requirements.txt
# Telegram & HTTP
python-telegram-bot>=21.0,<23
python-dotenv>=1.0.0
aiohttp>=3.9.0

# Descargas (libros / audio / video)
yt-dlp>=2024.1.0

# Tests
pytest>=7.4.0
